+++ __mocks__/react-dnd-scrollzone.js
@@ -1,0 +1,3 @@
+module.exports = el => el;
+module.exports.createVerticalStrength = () => {};
+module.exports.createHorizontalStrength = () => {};