+++ examples/storybooks/external-node.js
@@ -1,70 +1,79 @@
+/* eslint-disable react/no-multi-comp */
 import React, { Component } from 'react';
-import { DragDropContext } from 'react-dnd';
+import PropTypes from 'prop-types';
+import { DragDropContext, DragSource } from 'react-dnd';
 import HTML5Backend from 'react-dnd-html5-backend';
-import {
-  SortableTreeWithoutDndContext as SortableTree,
-  insertNode,
-  dndWrapExternalSource,
-} from '../../src';
+import { SortableTreeWithoutDndContext as SortableTree } from '../../src';
 
+// -------------------------
+// Create an drag source component that can be dragged into the tree
+// https://react-dnd.github.io/react-dnd/docs-drag-source.html
+// -------------------------
+// This type must be assigned to the tree via the `dndType` prop as well
 const externalNodeType = 'yourNodeType';
+const externalNodeSpec = {
+  // This needs to return an object with a property `node` in it.
+  // Object rest spread is recommended to avoid side effects of
+  // referencing the same object in different trees.
+  beginDrag: componentProps => ({ node: { ...componentProps.node } }),
+};
+const externalNodeCollect = (connect /* , monitor */) => ({
+  connectDragSource: connect.dragSource(),
+  // Add props via react-dnd APIs to enable more visual
+  // customization of your component
+  // isDragging: monitor.isDragging(),
+  // didDrop: monitor.didDrop(),
+});
+class externalNodeBaseComponent extends Component {
+  render() {
+    const { connectDragSource, node } = this.props;
 
-// this will wrap your external node component as a valid react-dnd DragSource
-const YourExternalNodeComponent = dndWrapExternalSource(
-  ({ node }) =>
-    <div
-      style={{
-        border: 'solid black 1px',
-        display: 'inline-block',
-        padding: '3px 5px',
-        background: 'blue',
-        color: 'white',
-      }}
-    >
-      {node.title}
-    </div>,
-  externalNodeType
-);
+    return connectDragSource(
+      <div
+        style={{
+          display: 'inline-block',
+          padding: '3px 5px',
+          background: 'blue',
+          color: 'white',
+        }}
+      >
+        {node.title}
+      </div>,
+      { dropEffect: 'copy' }
+    );
+  }
+}
+externalNodeBaseComponent.propTypes = {
+  node: PropTypes.shape({ title: PropTypes.string }).isRequired,
+  connectDragSource: PropTypes.func.isRequired,
+};
+const YourExternalNodeComponent = DragSource(
+  externalNodeType,
+  externalNodeSpec,
+  externalNodeCollect
+)(externalNodeBaseComponent);
 
 class App extends Component {
   constructor(props) {
     super(props);
 
     this.state = {
-      treeData: [{ title: 'node1' }, { title: 'node2' }],
+      treeData: [{ title: 'Mama Rabbit' }, { title: 'Papa Rabbit' }],
     };
   }
 
   render() {
     return (
       <div>
-        <div style={{ height: 250 }}>
+        <div style={{ height: 300 }}>
           <SortableTree
             treeData={this.state.treeData}
             onChange={treeData => this.setState({ treeData })}
             dndType={externalNodeType}
           />
         </div>
-        <YourExternalNodeComponent
-          node={{ title: 'External node' }}
-          addNewItem={newItem => {
-            const { treeData } = insertNode({
-              treeData: this.state.treeData,
-              newNode: newItem.node,
-              depth: newItem.depth,
-              minimumTreeIndex: newItem.minimumTreeIndex,
-              expandParent: true,
-              getNodeKey: ({ treeIndex }) => treeIndex,
-            });
-            this.setState({ treeData });
-          }}
-          // Update the tree appearance post-drag
-          dropCancelled={() =>
-            this.setState(state => ({
-              treeData: state.treeData.concat(),
-            }))}
-        />
-        ↑ drag this
+        <YourExternalNodeComponent node={{ title: 'Baby Rabbit' }} />
+        ← drag this
       </div>
     );
   }